Initial Import
This is the initial import of the status light for the Build Bot builds

This system drives a consumer RGB LED using a [IR Toy](http://dangerousprototypes.com/docs/USB_IR_Toy_v2) USB-Serial device that can receive and transmit IR signals.

This code has a learning mode to facilitiate remotes that are different to the one initially used.